add/update docstrings across the assistant surface, make sure all types are exported, move ChatStream file into the assistant folder
