+++ src/assistant/chatStream.ts
@@ -1,6 +1,14 @@
 import { Readable } from 'stream';
-import { convertKeysToCamelCase } from './convertKeys';
+import { convertKeysToCamelCase } from '../utils/convertKeys';
 
+/**
+ * Implements an async iterable that processes the readable stream of an assistant chat response.
+ *
+ * This class expects each chunk of data in the stream to begin with `data:` and be followed by a valid chunk of JSON.
+ * If a chunk contains malformed JSON, it is skipped, and a debug message is logged.
+ *
+ * @template Item - The type of items yielded by the iterable.
+ */
 export class ChatStream<Item> implements AsyncIterable<Item> {
   private stream: Readable;
 
@@ -28,10 +36,6 @@ export class ChatStream<Item> implements AsyncIterable<Item> {
             console.debug(`Skipping malformed JSON:${line}`);
             continue;
           }
-        } else {
-          // skip malformed chunk
-          console.debug(`Skipping malformed chunk:${line}`);
-          continue;
         }
       }
     }
