+++ src/assistant/index.ts
@@ -26,6 +26,16 @@ export type {
 } from './control/types';
 export type {
   ChatOptions,
+  ChatModelEnum,
+  ChoiceModel,
+  FinishReasonEnum,
+  StreamedChatResponse,
+  StreamedChatCompletionResponse,
+  BaseChunk,
+  MessageStartChunk,
+  ContentChunk,
+  CitationChunk,
+  MessageEndChunk,
   ContextOptions,
   ListFilesOptions,
   UploadFileOptions,
@@ -33,18 +43,19 @@ export type {
   AssistantFileStatusEnum,
   AssistantFilesList,
 } from './data/types';
+export { ChatStream } from './chatStream';
 
 /**
  * The `Assistant` class holds the data plane methods for interacting with
  *  [Assistants](https://docs.pinecone.io/guides/assistant/understanding-assistant).
  *
- *  This class's methods are instantiated on a Pinecone object and are used to interact with the data plane of an Assistant.
+ *  This class can be instantiated through a {@link Pinecone} object, and is used to interact with a specific assistant.
  *
  *  @example
  *  ```typescript
  *  import { Pinecone } from '@pinecone-database/pinecone';
  *  const pc = new Pinecone();
- *  const assistant = pc.Assistant('assistant-name');
+ *  const assistant = pc.assistant('assistant-name');
  *  ```
  */
 export class Assistant {
@@ -65,7 +76,7 @@ export class Assistant {
   /**
    * Creates an instance of the `Assistant` class.
    *
-   * @param assistantName - The name of the Assistant.
+   * @param assistantName - The name of the assistant.
    * @param config - The Pinecone configuration object containing an API key and other configuration parameters
    * needed for API calls.
    *
@@ -117,14 +128,14 @@ export class Assistant {
   // --------- Chat methods ---------
 
   /**
-   * Sends a message to the Assistant and receives a response. Retries the request if the server fails.
+   * Sends a message to the assistant and receives a response. Retries the request if the server fails.
    *
    * @example
    * ```typescript
    * import { Pinecone } from '@pinecone-database/pinecone';
    * const pc = new Pinecone();
    * const assistantName = 'test1';
-   * const assistant = pc.Assistant(assistantName);
+   * const assistant = pc.assistant(assistantName);
    * const chatResp = await assistant.chat({messages: [{role: 'user', content: "What is the capital of France?"}]});
    * // {
    * //  id: '000000000000000023e7fb015be9d0ad',
@@ -136,71 +147,115 @@ export class Assistant {
    * //  model: 'gpt-4o-2024-05-13',
    * //  citations: [ { position: 209, references: [Array] } ],
    * //  usage: { promptTokens: 493, completionTokens: 38, totalTokens: 531 }
-   * }
+   * // }
    * ```
    *
    * @param options - A {@link ChatOptions} object containing the message and optional parameters to send to the
-   * Assistant.
-   * @returns A promise that resolves to a {@link ChatModel} object containing the response from the Assistant.
+   * assistant.
+   * @returns A promise that resolves to a {@link ChatModel} object containing the response from the assistant.
    */
   chat(options: ChatOptions) {
     return this._chat(options);
   }
 
   /**
-   * Sends a message to the Assistant and receives a streamed response. Retries the request if the server fails.
+   * Sends a message to the assistant and receives a streamed response as {@link ChatStream<StreamedChatResponse>}. Retries the request if the server fails.
    *
    * @example
    * ```typescript
    * import { Pinecone } from '@pinecone-database/pinecone';
    * const pc = new Pinecone();
    * const assistantName = 'test1';
-   * const assistant = pc.Assistant(assistantName);
-   * const chatResp = await assistant.chat({messages: [{role: 'user', content: "What is the capital of France?"}]});
-   * // {
-   * //  id: '000000000000000023e7fb015be9d0ad',
-   * //  finishReason: 'stop',
-   * //  message: {
-   * //    role: 'assistant',
-   * //    content: 'The capital of France is Paris.'
-   * //  },
-   * //  model: 'gpt-4o-2024-05-13',
-   * //  citations: [ { position: 209, references: [Array] } ],
-   * //  usage: { promptTokens: 493, completionTokens: 38, totalTokens: 531 }
+   * const assistant = pc.assistant(assistantName);
+   * const chatStream = await assistant.chatStream({ messages: [{ role: 'user', content: 'What is the capital of France?'}]});
+   *
+   * // stream the response and log each chunk
+   * for await (const chunk of newStream) {
+   *   console.log(chunk);
    * }
+   * // each chunk will have a variable shape depending on the type:
+   * // { type:"message_start", id:"response_id", model:"gpt-4o-2024-05-13", role:"assistant"}
+   * // { type:"content_chunk", id:"response_id", model:"gpt-4o-2024-05-13", delta:{ content:"The"}}
+   * // { type:"content_chunk", id:"response_id", model:"gpt-4o-2024-05-13", delta:{ content:" test"}}
+   * // { type:"message_end", id:"response_id", model:"gpt-4o-2024-05-13", finishReason:"stop",usage:{ promptTokens:371,completionTokens:48,totalTokens:419}}
    * ```
    *
    * @param options - A {@link ChatOptions} object containing the message and optional parameters to send to the
-   * Assistant.
-   * @returns A promise that resolves to a {@link ChatModel} object containing the response from the Assistant.
+   * assistant.
+   * @returns A promise that resolves to a {@link ChatStream<StreamedChatResponse>}.
    */
   chatStream(options: ChatOptions) {
     return this._chatStream(options);
   }
 
   /**
-   * Sends a message to the Assistant and receives a response. Response is compatible with
+   * Sends a message to the assistant and receives a response that is compatible with
    * [OpenAI's Chat Completion API](https://platform.openai.com/docs/guides/text-generation. Retries the request if the server fails.
    *
-   * See {@link chat} for example usage.
+   * @example
+   * ```typescript
+   * import { Pinecone } from '@pinecone-database/pinecone';
+   * const pc = new Pinecone();
+   * const assistantName = 'test1';
+   * const assistant = pc.assistant(assistantName);
+   * const chatCompletion = await assistant.chatCompletion({ messages: [{ role: 'user', content: 'What is the capital of France?' }]});
+   * console.log(chatCompletion);
+   * // {
+   * //  id: "response_id",
+   * //  choices: [
+   * //  {
+   * //    finishReason: "stop",
+   * //    index: 0,
+   * //    message: {
+   * //      role: "assistant",
+   * //      content: "The data mentioned is described as \"some temporary data\"  [1].\n\nReferences:\n1. [test-chat.txt](https://storage.googleapis.com/knowledge-prod-files/0b3dc744-fc03-4752-af7a-0fa25e2ea732%2F940864ff-f412-4828-b772-3e39442aa011%2F7d031d41-60b1-4517-9050-ba0e210de17f.txt?X-Goog-Algorithm=GOOG4-RSA-SHA256&X-Goog-Credential=ke-prod-1%40pc-knowledge-prod.iam.gserviceaccount.com%2F20250214%2Fauto%2Fstorage%2Fgoog4_request&X-Goog-Date=20250214T062103Z&X-Goog-Expires=3600&X-Goog-SignedHeaders=host&response-content-disposition=inline&response-content-type=text%2Fplain&X-Goog-Signature=54a609a124be5223d754bc7259a8307a5101f65c92680c98978ed23774e3eb22f53927bc18f38070167a670956b91419648c28d542fc97a8070851f4f27b6b492b621790858c6fa3c03986d8bb0c919348574f3539c503bac2b4879b026cac28298eb7dc7ebea3bbf61f9058eda9d04978a8bdd4dabd10f738d7d8b402f4ae517a85a42f5ce64ca4a8846789f3a45c296422898c638c677fe1d8035c135ee4a8496027da4fecc515990e532a54a9bad5f17a59a584c6092ba61c92f92eccc308e09df40150da55f1317f0e9a9e38f88f19ce5bbe6480aaaf265d4067b641bfdc0f425d7b2dd8ceb007f05c1847e60cb17b76205a804b13d2395385233b18d55e) \n"
+   * //    }
+   * //   }
+   * //  ],
+   * //  model: "gpt-4o-2024-05-13",
+   * //  usage: {
+   * //    promptTokens: 371,
+   * //    completionTokens: 19,
+   * //    totalTokens: 390
+   * //  }
+   * // }
+   * ```
    *
    * @param options - A {@link ChatOptions} object containing the message and optional parameters to send
-   * to an Assistant.
-   * @returns A promise that resolves to a {@link ChatCompletionModel} object containing the response from the Assistant.
+   * to an assistant.
+   * @returns A promise that resolves to a {@link ChatCompletionModel} object containing the response from the assistant.
    */
   chatCompletion(options: ChatOptions) {
     return this._chatCompletion(options);
   }
 
   /**
-   * Sends a message to the Assistant and receives a streamed response. Response is compatible with
+   * Sends a message to the assistant and receives a streamed response as {@link ChatStream<StreamedChatCompletionResponse>}. Response is compatible with
    * [OpenAI's Chat Completion API](https://platform.openai.com/docs/guides/text-generation. Retries the request if the server fails.
    *
-   * See {@link chat} for example usage.
+   * @example
+   * ```typescript
+   * import { Pinecone } from '@pinecone-database/pinecone';
+   * const pc = new Pinecone();
+   * const assistantName = 'test1';
+   * const assistant = pc.assistant(assistantName);
+   * const chatStream = await assistant.chatCompletionStream({messages: [{role: 'user', content: "What is the capital of France?"}]});
+   *
+   * // stream the response and log each chunk
+   * for await (const chunk of newStream) {
+   *   if (chunk.choices.length > 0 && chunk.choices[0].delta.content) {
+   *     process.stdout.write(chunk.choices[0].delta.content);
+   *   }
+   * }
+   * // { id: 'response_id', choices: [{ index: 0, delta: { role: 'assistant' }, finishReason: null }], model: 'gpt-4o-2024-05-13', usage: null }
+   * // { id: 'response_id', choices: [{ index: 0, delta: { content: 'The' }}, finishReason: null }], model: 'gpt-4o-2024-05-13', usage: null }
+   * // { id: 'response_id', choices: [{ index: 0, delta: { content: ' test' }}, finishReason: null }], model: 'gpt-4o-2024-05-13', usage: null }
+   * // { id: 'response_id', choices: [], model: 'gpt-4o-2024-05-13', usage: { promptTokens: 371, completionTokens: 48, totalTokens: 419 }}
+   * ```
    *
    * @param options - A {@link ChatOptions} object containing the message and optional parameters to send
-   * to an Assistant.
-   * @returns A promise that resolves to a {@link ChatCompletionModel} object containing the response from the Assistant.
+   * to an assistant.
+   * @returns A promise that resolves to a {@link ChatStream<StreamedChatCompletionResponse>}.
    */
   chatCompletionStream(options: ChatOptions) {
     return this._chatCompletionStream(options);
@@ -209,14 +264,14 @@ export class Assistant {
   // --------- File methods ---------
 
   /**
-   * Lists files (with optional filter) uploaded to an Assistant.
+   * Lists files (with optional filter) uploaded to an assistant.
    *
    * @example
    * ```typescript
    * import { Pinecone } from '@pinecone-database/pinecone';
    * const pc = new Pinecone();
    * const assistantName = 'test1';
-   * const assistant = pc.Assistant(assistantName);
+   * const assistant = pc.assistant(assistantName);
    * const files = await assistant.listFiles({filter: {metadata: {key: 'value'}}});
    * console.log(files);
    * // {
@@ -247,14 +302,14 @@ export class Assistant {
   }
 
   /**
-   * Describes a file uploaded to an Assistant.
+   * Describes a file uploaded to an assistant.
    *
    * @example
    * ```typescript
    * import { Pinecone } from '@pinecone-database/pinecone';
    * const pc = new Pinecone();
    * const assistantName = 'test1';
-   * const assistant = pc.Assistant(assistantName);
+   * const assistant = pc.assistant(assistantName);
    * const files = await assistant.listFiles();
    * let fileId: string;
    * if (files.files) {
@@ -285,7 +340,7 @@ export class Assistant {
   }
 
   /**
-   * Uploads a file to an Assistant.
+   * Uploads a file to an assistant.
    *
    * Note: This method does *not* use the generated code from the OpenAPI spec.
    *
@@ -294,7 +349,7 @@ export class Assistant {
    * import { Pinecone } from '@pinecone-database/pinecone';
    * const pc = new Pinecone();
    * const assistantName = 'test1';
-   * const assistant = pc.Assistant(assistantName);
+   * const assistant = pc.assistant(assistantName);
    * await assistant.uploadFile({path: "test-file.txt", metadata: {"test-key": "test-value"}})
    * // {
    * //  name: 'test-file.txt',
@@ -317,14 +372,14 @@ export class Assistant {
   }
 
   /**
-   * Deletes a file uploaded to an Assistant by ID.
+   * Deletes a file uploaded to an assistant by ID.
    *
    * @example
    * ```typescript
    * import { Pinecone } from '@pinecone-database/pinecone';
    * const pc = new Pinecone();
    * const assistantName = 'test1';
-   * const assistant = pc.Assistant(assistantName);
+   * const assistant = pc.assistant(assistantName);
    * const files = await assistant.listFiles();
    * let fileId: string;
    * if (files.files) {
@@ -342,14 +397,14 @@ export class Assistant {
 
   /**
    * Retrieves [the context snippets](https://docs.pinecone.io/guides/assistant/understanding-context-snippets) used
-   * by an Assistant during the retrieval process.
+   * by an assistant during the retrieval process.
    *
    * @example
    * ```typescript
    * import { Pinecone } from '@pinecone-database/pinecone';
    * const pc = new Pinecone();
    * const assistantName = 'test1';
-   * const assistant = pc.Assistant(assistantName);
+   * const assistant = pc.assistant(assistantName);
    * const response = await assistant.context({query: "What is the capital of France?"});
    * console.log(response);
    * // {
