+++ src/utils/index.ts
@@ -4,7 +4,7 @@ import { queryParamsStringify } from './queryParamsStringify';
 import { buildUserAgent } from './user-agent';
 import { getFetch } from './fetch';
 import { RetryOnServerFailure } from './retries';
-import { ChatStream } from './chatStream';
+import { ChatStream } from '../assistant/chatStream';
 import { convertKeysToCamelCase } from './convertKeys';
 
 export {
