+++ ts-external-app-test/chatStreamTest.ts
@@ -3,21 +3,17 @@ import { Pinecone } from '../dist/index';
 const client = new Pinecone();
 
 const newStream = await client
-  .assistant('test-chat-stream')
-  .chatCompletionStream({ messages: ['tell me about API design'] });
+  .assistant('test-chat')
+  .chatStream({ messages: ['tell me about API design'] });
 
 (async () => {
-  // print entire chunks
   for await (const chunk of newStream) {
-    if (chunk.choices.length > 0 && chunk.choices[0].delta.content) {
-      process.stdout.write(chunk.choices[0].delta.content);
-    }
+    console.log(JSON.stringify(chunk));
   }
+})();
 
-  // // print only content
-  // for await (const chunk of newStream) {
-  //   if (chunk.type === 'content_chunk' && chunk.delta?.content) {
-  //     process.stdout.write(chunk.delta.content);
-  //   }
-  // }
+(async () => {
+  for await (const chunk of newStream) {
+    console.log(JSON.stringify(chunk));
+  }
 })();
